[REFACTOR] - Extract shared formatters and GRID constant; clean up row component
- Move formatPhone/formatPaymentDate to lib/utils.ts (were duplicated in member-payment-row and payment-status-grid)
- Export GRID template string from member-payment-row; import in sortable-payment-table to prevent column drift
- Remove unnecessary fragment wrapper from MemberPaymentRow
- Simplify onClick={() => onSelect()} to onClick={onSelect}
- Remove WHAT-explaining inline comments from row grid cells
